+++ starsky/starsky/clientapp/src/components/organisms/modal-detailview-rename-file/modal-detailview-rename-file.tsx
@@ -7,6 +7,7 @@ import { IExifStatus } from '../../../interfaces/IExifStatus';
 import { newIFileIndexItem } from '../../../interfaces/IFileIndexItem';
 import FetchPost from '../../../shared/fetch-post';
 import { FileExtensions } from '../../../shared/file-extensions';
+import { FileListCache } from '../../../shared/filelist-cache';
 import { Language } from '../../../shared/language';
 import { UrlQuery } from '../../../shared/url-query';
 import FormControl from '../../atoms/form-control/form-control';
@@ -22,9 +23,10 @@ const ModalDetailviewRenameFile: React.FunctionComponent<IModalRenameFileProps> 
   // content
   const settings = useGlobalSettings();
   const language = new Language(settings.language);
-  const MessageNonValidExtension: string = language.text("Dit bestand kan zo niet worden weggeschreven", "This file cannot be saved");
-  const MessageChangeToDifferentExtension = language.text("Let op! Je veranderd de extensie van het bestand, deze kan hierdoor onleesbaar worden",
-    "Pay attention! You change the file extension, which can make it unreadable");
+  const MessageNonValidExtension: string = language.text("Dit bestand kan zo niet worden weggeschreven",
+    "This file cannot be saved");
+  const MessageChangeToDifferentExtension = language.text("Let op! Je veranderd de extensie van het bestand, "
+    + "deze kan hierdoor onleesbaar worden", "Pay attention! You change the file extension, which can make it unreadable");
   const MessageGeneralError: string = language.text("Er is iets misgegaan met de aanvraag, probeer het later opnieuw",
     "Something went wrong with the request, please try again later");
   const MessageRenameFileName = language.text("Bestandsnaam wijzigen", "Rename file name");
@@ -117,6 +119,7 @@ const ModalDetailviewRenameFile: React.FunctionComponent<IModalRenameFileProps> 
       return
     }
 
+    new FileListCache().CacheCleanEverything();
     // redirect to new path (so if you press refresh the image is shown)
     const replacePath = new UrlQuery().updateFilePathHash(history.location.search, filePathAfterChange);
     await history.navigate(replacePath, { replace: true });