+++ starsky/starsky/clientapp/src/index.tsx
@@ -1,4 +1,6 @@
 import 'abortcontroller-polyfill/dist/abortcontroller-polyfill-only'; // for the feature
+import 'core-js/features/array/filter'; // array filter
+import 'core-js/features/array/some'; // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/some
 import 'core-js/features/dom-collections/for-each'; // queryselector.forEach
 import 'core-js/features/object'; // Object.entries is not a function
 import 'core-js/features/promise'; // Yes I promise