+++ history.md
@@ -35,6 +35,7 @@ node starsky-tools/build-tools/app-version-update.js
 # version 0.4.0-beta.0 _(Unreleased)_  2020-10-??
 _In this release websockets are used_
 - [x]   (Added) _Front-end_ Update view when other clients are updating content
+- [x]   (Changed) _Front-end_ In GPX view mode & when unlocked: touchZoom and doubleClickZoom are enabled
 - [x]   (Fixed) _Front-end_ When file is added to view, the colorClassActiveList is updated
 - [x]   (Fixed) _Front-end_ When folder or file is renamed the clientside cache is not correct
 - [x]   (Fixed) _Front-end_ When in Archive mode and 'Move file to trash' client cache is cleared