+++ history.md
@@ -33,7 +33,7 @@ node starsky-tools/build-tools/app-version-update.js
 - [ ]   (x) Move multiple files __not implemented__
 
 # version 0.4.0-beta.0 _(Unreleased)_  2020-10-??
-_In this release websockets are used_
+_New Feature: In this release websockets are used (note: when using reverse config)_
 - [x]   (Added) _Front-end_ Update view when other clients are updating content
 - [x]   (Changed) _Front-end_ In GPX view mode & when unlocked: touchZoom and doubleClickZoom are enabled
 - [x]   (Fixed) _Front-end_ When file is added to view, the colorClassActiveList is updated